server-app/requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
bcrypt==4.2.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.67.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
packaging==24.2
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.37
tqdm==4.66.5
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
